[glsl-new] Refactor type handling
